+++ packages/suite/src/components/chat/components/ChatInput.tsx
@@ -1,12 +1,10 @@
-import { cn } from '@/lib/utils';
-import { border } from '@/styles/theme';
 import { Textarea } from '@/components/ui/textarea';
+import { cn } from '@/lib/utils';
+import { border, pressable, text } from '@/styles/theme';
 import { Button } from '@/components/ui/button';
-import { Loader2, Send } from 'lucide-react';
-import { useSuite } from '@/hooks/use-suite';
-import { LazyAnimatedBuster } from '@getgrowly/ui';
-import { text, pressable } from '@/styles/theme';
+import { Loader2, LucideSend } from 'lucide-react';
 import { useSuiteSession } from '@/hooks/use-session';
+import { useThemeStyles } from '@/hooks/use-theme-styles';
 
 export const ChatInput = ({
   sendMessageHandler,
@@ -15,8 +13,8 @@ export const ChatInput = ({
   sendMessageHandler: () => void;
   isSending: boolean;
 }) => {
-  const { config } = useSuite();
-  const { busterState, setBusterState, inputValue, setInputValue } = useSuiteSession();
+  const styles = useThemeStyles();
+  const { inputValue, setInputValue } = useSuiteSession();
 
   // If "ENTER" is clicked, send a message.
   const handleKeyDown = (event: React.KeyboardEvent<HTMLTextAreaElement>) => {
@@ -28,11 +26,8 @@ export const ChatInput = ({
   };
 
   return (
-    <div
-      className={cn('p-4 border-t', border.lineDefault)}
-      style={{ backgroundColor: config?.theme?.background }}>
+    <div className={cn('p-4 border-t', border.lineDefault)} style={styles.chat.input}>
       <div className={cn('flex space-x-2', text.body)}>
-        <LazyAnimatedBuster state={busterState} setState={setBusterState} width={40} height={40} />
         <Textarea
           value={inputValue}
           onChange={e => setInputValue(e.target.value)}
@@ -41,24 +36,30 @@ export const ChatInput = ({
           placeholder="Send a message..."
           style={{
             border: 'none',
+            backgroundColor: styles.chat.input.backgroundColor,
+            color: styles.chat.input.color,
           }}
           className={cn(
             'flex-1',
             border.lineDefault,
-            'placeholder:text-gray-500 text-xs placeholder:text-xs focus:outline-none focus:ring-0'
+            text.body,
+            'placeholder:text-gray-500 text-sm placeholder:text-sm focus:outline-none focus:ring-0'
           )}
         />
         <Button
           className={cn(border.defaultActive, pressable.inverse, text.headline)}
           style={{
-            backgroundColor: config?.theme?.primary,
-            color: config?.theme?.text,
+            color: styles.text.inverse.color,
             borderRadius: '50%',
             width: '40px',
             height: '40px',
           }}
           onClick={sendMessageHandler}>
-          {isSending ? <Loader2 className="h-4 w-4 animate-spin" /> : <Send className="h-4 w-4" />}
+          {isSending ? (
+            <Loader2 className="h-4 w-4 animate-spin" />
+          ) : (
+            <LucideSend className="h-4 w-4" />
+          )}
         </Button>
       </div>
     </div>