+++ packages/suite/src/components/panel/components/PanelLayout.tsx
@@ -6,7 +6,7 @@ export const PanelLayout = ({ children }: { children: React.ReactNode }) => {
   const { config } = useSuite();
   return (
     <ScrollArea
-      className={cn('flex-1', config?.display === 'fullView' ? 'max-h-[90vh]' : 'max-h-[500px]')}
+      className={cn('flex-1', config?.display === 'fullView' ? 'max-h-[90vh]' : 'max-h-[70vh]')}
       style={{ padding: '0px 50px' }}
       id="chat-messages-container">
       {children}