+++ packages/suite/src/components/providers/SuiteProvider.tsx
@@ -1,15 +1,17 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { OnchainKitProvider } from '@coinbase/onchainkit';
 import { useSuiteSession } from '../../hooks/use-session';
-import { Loader2 } from 'lucide-react';
 import { WalletConnectProvider } from './WalletConnectProvider';
 import { SuiteConfig, SuiteGlobalContext } from './SuiteProvider.types';
-import { Theme } from '../widgets';
+import { ThemeProvider } from './ThemeProvider';
+import { ThemeName } from '@/types/theme';
 import { WorkflowExecutionObserver } from './WorkflowExecutionObserver';
 
 const defaultConfig: SuiteConfig = {
   display: 'fullView',
-  theme: Theme.monoTheme,
+  themeMode: ThemeName.Light,
+  theme: {}, // Theme overrides will be handled by ThemeProvider
+  brandName: 'Cream',
 };
 
 export const SuiteContext = React.createContext<
@@ -36,7 +38,7 @@ export const SuiteProvider: React.FC<{
   children: React.ReactNode;
   context: SuiteGlobalContext;
 }> = ({ children, context }) => {
-  const [baseComponent, setBaseComponent] = useState<React.ReactNode>(<></>);
+  const [baseComponent, setBaseComponent] = useState<React.ReactNode>(<>children</>);
   const [isInitialized, setIsInitialized] = useState(false);
   const { createUserFromAddressIfNotExist, fetchOrganizationAgentById } = useSuiteSession();
   const [config, setConfig] = useState<SuiteConfig>(context.config ?? defaultConfig);
@@ -67,10 +69,7 @@ export const SuiteProvider: React.FC<{
   }, [context, walletAddress]);
 
   useEffect(() => {
-    let baseComponent = (
-      <>{isInitialized ? children : <Loader2 className="h-5 w-5 animate-spin" />}</>
-    );
-
+    let baseComponent = children;
     if (!walletAddress && !context.session?.connect) {
       console.log('Growly Suite: Wallet is not connected');
     }
@@ -104,7 +103,11 @@ export const SuiteProvider: React.FC<{
           setWalletAddress,
         },
       }}>
-      <WorkflowExecutionObserver>{baseComponent}</WorkflowExecutionObserver>
+      <ThemeProvider
+        defaultTheme={config.themeMode || ThemeName.Light}
+        themeOverrides={config.theme}>
+        <WorkflowExecutionObserver>{baseComponent}</WorkflowExecutionObserver>
+      </ThemeProvider>
     </SuiteContext.Provider>
   );
 };