+++ packages/suite/src/components/widgets/index.ts
@@ -5,4 +5,3 @@ export * from './components/ChatWidget';
 export * from './components/DemoChatWidget';
 export * from './components/StaticWidget';
 export * from './components/GrowlyButton';
-export * as Theme from './theme';