+++ packages/website/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @getgrowly/website
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @getgrowly/suite@1.0.5
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @getgrowly/suite@1.0.4
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @getgrowly/suite@1.0.3
+
 ## 0.1.4
 
 ### Patch Changes